fastapi_model/requirements.txt
fastapi
uvicorn[standard]
tensorflow
pillow
numpy
python-multipart
mangum
